fix(coinmarket): receive currency in modal
